+++ src/Exceptions/FileNotFoundException.php
@@ -14,7 +14,7 @@ use Throwable;
  */
 class FileNotFoundException extends Exception
 {
-    public function __construct($path, $code = 0, Throwable $previous = null)
+    public function __construct($path, $code = 0, ?Throwable $previous = null)
     {
         parent::__construct(sprintf('The file "%s" was not found.', $path), $code, $previous);
     }