+++ tests/TestCase.php
@@ -8,8 +8,10 @@ use Illuminate\Contracts\Filesystem\Filesystem;
 use Illuminate\Support\Facades\Storage;
 use Orchestra\Testbench\TestCase as OrchestraTestCase;
 use ReflectionClass;
+use ReflectionException;
 use ReflectionMethod;
 use ReflectionProperty;
+use function Orchestra\Testbench\package_path;
 
 class TestCase extends OrchestraTestCase
 {
@@ -36,14 +38,19 @@ class TestCase extends OrchestraTestCase
         ];
     }
 
+    protected function getApplicationBasePath(): string
+    {
+        return package_path() . '/vendor/orchestra/testbench-core/laravel/';
+    }
+
+    /**
+     * @throws ReflectionException
+     */
     protected function getAccessibleReflectionMethod(string $method): ReflectionMethod
     {
         $reflectionProtector = new ReflectionClass($this->protector);
-        $method = $reflectionProtector->getMethod($method);
 
-        $method->setAccessible(true);
-
-        return $method;
+        return $reflectionProtector->getMethod($method);
     }
 
     /**
@@ -52,6 +59,7 @@ class TestCase extends OrchestraTestCase
      * @param string $methodName
      * @param array $params
      * @return mixed
+     * @throws ReflectionException
      */
     protected function runProtectedMethod(string $methodName, array $params = []): mixed
     {
@@ -60,11 +68,13 @@ class TestCase extends OrchestraTestCase
         return $method->invoke($this->protector, ...$params);
     }
 
+    /**
+     * @throws ReflectionException
+     */
     protected function setProtectedProperty(string $propertyName, mixed $value): void
     {
         $property = new ReflectionProperty($this->protector, $propertyName);
 
-        $property->setAccessible(true);
         $property->setValue($this->protector, $value);
     }